PD#134530: di: update txt mode and disable dm for 1080 input
1.update di txt mode
2.disable dm for 1080 input base on hw limitation
3.merge from iptv base on Bug132938 and Bug132509

+++ drivers/amlogic/deinterlace/dnr.h
@@ -93,6 +93,6 @@ int ver_blk_ofst_calc_sw(int *pVbOfVldCnt,
 			 int prm_vbof_vldcntthd,
 			 int nRow,
 			 int nCol);
-void run_dnr_in_irq(int nCol, int nRow);
+void run_dnr_in_irq(unsigned short nCol, unsigned short nRow);
 #endif
